Move waitForAll types test to new file
