+++ tests/types.test.tsx
@@ -1,26 +1,7 @@
 import { expectType } from 'ts-expect'
 import type { Atom, PrimitiveAtom, WritableAtom } from 'jotai'
 import { atom, useAtom } from 'jotai'
-import { useAtomValue, waitForAll } from 'jotai/utils'
 import type { SetAtom } from '../src/core/atom'
-
-it('waitForAll() should return the correct types', () => {
-  function Component() {
-    const numberAtom = atom(async () => 0)
-    const stringAtom = atom(async () => '')
-
-    const [number, string] = useAtomValue(waitForAll([numberAtom, stringAtom]))
-    expectType<number>(number)
-    expectType<string>(string)
-
-    const { numberAtom: number2, stringAtom: string2 } = useAtomValue(
-      waitForAll({ numberAtom, stringAtom })
-    )
-    expectType<number>(number2)
-    expectType<string>(string2)
-  }
-  Component
-})
 
 it('atom() should return the correct types', () => {
   function Component() {
