+++ tests/utils/waitForAll.types.test.tsx
@@ -1,0 +1,21 @@
+import { expectType } from 'ts-expect'
+import { atom } from 'jotai'
+import { useAtomValue, waitForAll } from 'jotai/utils'
+
+it('waitForAll() should return the correct types', () => {
+  function Component() {
+    const numberAtom = atom(async () => 0)
+    const stringAtom = atom(async () => '')
+
+    const [number, string] = useAtomValue(waitForAll([numberAtom, stringAtom]))
+    expectType<number>(number)
+    expectType<string>(string)
+
+    const { numberAtom: number2, stringAtom: string2 } = useAtomValue(
+      waitForAll({ numberAtom, stringAtom })
+    )
+    expectType<number>(number2)
+    expectType<string>(string2)
+  }
+  Component
+})
